Create polygon.md
<b>Polygon</b>

Why did we choose Polygon? The Polygon network is carbon free - since it uses the PoS (Proof of Stake) system - and is ideal for Blockchain projects compatible with the Ethereum network, the main one when it comes to smart contracts.

The Polygon network also gives flexibility to the Greener project with the creation of tokens linked to the Carbon Storage Unit (UEC), within the ERC-777 protocol. In this way, the traceability of the token can be checked at any time.

As an option, the user can purchase GPTs tokens (Greener Protection Tokens) and pay them with Matic, the cryptocurrency of the Polygon network, where gas fees are more accessible and enable secure transactions through a digital wallet.

<b>What is Polygon?</b>

Polygon is a protocol that allows building and connecting Blockchain networks compatible with the Ethereum network (EVM), enabling numerous applications and the development of smart contracts for customizable solutions.

Founded in 2017 as Matic Network, Polygon has gained popularity and its coin, MATIC, is one of the top 15 in the crypto market.

Primarily, Ethereum focused on decentralization and security, leaving aside scalability, making its transactions very expensive and slow. Polygon came to solve this problem with faster and cheaper transactions.

As an Ethereum-derived protocol, Polygon stands out for improved transactions and lower implementation costs for developers.

To give you an idea, today Polygon can process up to 7,000 transactions per second (TPS), while Ethereum can only handle 14 transactions per second. In other words, Polygon's speed reduces costs and is more efficient.

<b>What is MATIC?</b>

MATIC is the cryptocurrency of the Polygon network and serves for trading and transactions within the Polygon network.

MATIC can be defined as a utility token within the Polygon framework, in the ERC-20 protocol.

In the Greener project, MATIC is used as a currency for purchasing the GPT token, which can be done through the Greener platform (www.begreener.io).

The use of MATIC brings security and reliability to the process as it is one of the main cryptocurrencies available. To give you an idea, Polygon today houses more than 19,000 decentralized applications.

<b>Proof-of-Stake Model</b>

The proof-of-stake consensus mechanism remunerates token holders by keeping the network operational and verifying transactions.

Proof-of-Stake relies on its operators to seek the reward of the action, reducing costs and being more practically carbon neutral, without the need to process a transaction in Ethereum, for example.

In short, proof-of-stake is an alternative to the proof-of-work (PoW) model where validating agents use their own cryptocurrencies to validate actions. A node must have access to a certain amount of coins before being accepted by the network.